+++ src/views/business-details/business-details.njk
@@ -66,7 +66,7 @@
           Mobile: 07700 900 967
         </dd>
         <dd class="govuk-summary-list__actions">
-          <a class="govuk-link govuk-link--no-visited-state" href="/business-phone-change">Change<span class="govuk-visually-hidden"> business phone numbers</span></a>
+          <a class="govuk-link govuk-link--no-visited-state" href="/business-phone-numbers-change">Change<span class="govuk-visually-hidden"> business phone numbers</span></a>
         </dd>
       </div>
       <div class="govuk-summary-list__row">
