+++ src/views/business-details/business-phone-numbers-change.njk
@@ -1,0 +1,83 @@
+{% extends 'common/layout.njk' %}
+
+{% from "govuk/components/input/macro.njk" import govukInput %}
+{% from "govuk/components/button/macro.njk" import govukButton %}
+{% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
+{% from "common/navigation/back-link.njk" import appBackLink %}
+{% from "common/sub-header/macro.njk" import subHeader %}
+
+{% set pageTitle = "What are your business phone numbers?" %}
+{% set metaDescription = "Update the phone numbers for your business." %}
+
+{% block beforeContent %}
+    {{ subHeader(
+    businessName="Agile Farm Ltd", 
+    sbi="123456789", 
+    userName="Alfred Waldron"
+) }}
+
+    {{ appBackLink() }}
+
+{% endblock %}
+
+{% block content %}
+
+{% if errors %}
+    {{ govukErrorSummary({
+        titleText: "There is a problem",
+        errorList: errors | toErrorList
+    }) }}
+{% endif %}
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <form method="post" novalidate>
+      <fieldset class="govuk-fieldset">
+        <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
+          <h1 class="govuk-fieldset__heading">
+            {{ pageTitle }}
+          </h1>
+        </legend>
+
+        <div class="govuk-hint">
+          Enter at least one phone number
+        </div>
+
+        {{ govukInput({
+          label: {
+            text: "Business telephone number"
+          },
+          classes: "govuk-input--width-20",
+          id: "businessTelephone",
+          name: "businessTelephone",
+          value: addressPostcode,
+          autocomplete: "business-telephone",
+          errorMessage: errors.businessTelephone and {
+            text: errors.businessTelephone.text
+          }
+        }) }}
+
+        {{ govukInput({
+          label: {
+            text: "Business mobile number"
+          },
+          classes: "govuk-input--width-20",
+          id: "businessMobile",
+          name: "businessMobile",
+          value: addressPostcode,
+          autocomplete: "business-mobile",
+          errorMessage: errors.businessMobile and {
+            text: errors.businessMobile.text
+          }
+        }) }}
+      </fieldset>
+
+      {{ govukButton({
+        text: "Continue"
+      }) }}
+      
+      <p class="govuk-body"><a href="/business-details" class="govuk-link govuk-link--no-visited-state">Cancel</a></p>
+    </form>
+  </div>
+</div>
+{% endblock %}
